update on connection and CM

+++ src/dona/util/Statistic.java
@@ -7,9 +7,9 @@ import dona.entity.Message;
 import peersim.util.IncrementalStats;
 
 public class Statistic {
-	public static int query_index;
+	public static long query_index;
 	public static long total_time;
-	public static int query_complete;
+	public static long query_complete;
 	public static IncrementalStats is;
 
 }
